Observation.jl

MiniObserve.jl

Introduction

Observation.jl

StatsAccumulator.jl

MiniObserve.Observation — Module

This module implements the observation logic itself.

MiniObserve.Observation.joined_named_tuple_T — Method

concatenate named tuple and/or struct types into one single named tuple

MiniObserve.Observation.log_results — Method

Print results stored in stats to output using field separator FS and line separator LS.

MiniObserve.Observation.print_header — Method

Print a header for an observation type stats_t to output using field separator FS, name separator NS and line separator LS.

MiniObserve.Observation.stat_names — Method

translate accumulator types into prefixes for the header (e.g. min, max, etc.)

MiniObserve.Observation.to_named_tuple — Method

construct a named tuple from x

MiniObserve.Observation.tuple_type — Method

obtain a named tuple type with the same field types and names as struct_T

MiniObserve.Observation.@observe — Macro

@observe(statstype, model [, user_arg1...], declarations)

Generate a full analysis suite for a model.

Given a declaration

@observe Data model user1 user2 begin
	@record "time"      model.time
	@record "N"     Int length(model.population)

	@for ind in model.population begin
		@stat("capital", MaxMinAcc{Float64}, MeanVarAcc{FloatT}) <| ind.capital
		@stat("n_alone", CountAcc)           <| has_neighbours(ind)
	end

	@record u1			user1
	@record u2			user1 * user2
end

a type Data will be generated that provides (at least) the following members:

struct Data
	time :: Float64
	N :: Int
	capital :: @NamedTuple{max :: Float64, min :: Float64, mean :: Float64, var :: Float64}
	n_alone :: @NamedTuple{N :: Int}
	u1 :: Float64
	u2 :: Float64
end

The macro will also create a method for observe(::Type{Data), model...) that will perform the required calculations and returns a Data object. Use print_header to print a header for the generated type to an output and log_results to print the content of a data object.

Index

MiniObserve.Observation

MiniObserve.StatsAccumulator.CountAcc

MiniObserve.StatsAccumulator.HistAcc

MiniObserve.StatsAccumulator.HistAcc

MiniObserve.StatsAccumulator.MaxMinAcc

MiniObserve.StatsAccumulator.MeanVarAcc

MiniObserve.StatsAccumulator.SumAcc

MiniObserve.Observation.joined_named_tuple_T

MiniObserve.Observation.log_results

MiniObserve.Observation.print_header

MiniObserve.Observation.stat_names

MiniObserve.Observation.to_named_tuple

MiniObserve.Observation.tuple_type

MiniObserve.StatsAccumulator.result_type

MiniObserve.StatsAccumulator.results

MiniObserve.Observation.@observe