StatsAccumulator.jl

MiniObserve.jl

Introduction

Observation.jl

StatsAccumulator.jl

MiniObserve.StatsAccumulator.result_type — Method

result_type(_)

Returns the type of the results of accumulator T.

Per default the accumulator type itself is assumed to form the result. Overload for custom result types.

MiniObserve.StatsAccumulator.results — Method

results(t)

Return the results of an accumulator.

Per default it is assumed that the accumulator type contains the result. Overload for custom result types.

MiniObserve.StatsAccumulator.CountAcc — Type

Count number of true values in input. Results are returned as (;n).

MiniObserve.StatsAccumulator.HistAcc — Type

Track a histogram of input values. Results are returned as a (;bins::Vector{Int}).

MiniObserve.StatsAccumulator.HistAcc — Method

Construct a histogram accumulator. Provide minimum value and bin size as min and width.

MiniObserve.StatsAccumulator.MaxMinAcc — Type

Keep track of maximum and minimum of input. Results are returned as (;max, min).

MiniObserve.StatsAccumulator.MeanVarAcc — Type

Calculate mean and variance of input. Results are given as (;mean, var)

MiniObserve.StatsAccumulator.SumAcc — Type

Calculate sum of input. Results are returned as (;n).

Index

MiniObserve.Observation

MiniObserve.StatsAccumulator.CountAcc

MiniObserve.StatsAccumulator.HistAcc

MiniObserve.StatsAccumulator.HistAcc

MiniObserve.StatsAccumulator.MaxMinAcc

MiniObserve.StatsAccumulator.MeanVarAcc

MiniObserve.StatsAccumulator.SumAcc

MiniObserve.Observation.joined_named_tuple_T

MiniObserve.Observation.log_results

MiniObserve.Observation.print_header

MiniObserve.Observation.stat_names

MiniObserve.Observation.to_named_tuple

MiniObserve.Observation.tuple_type

MiniObserve.StatsAccumulator.result_type

MiniObserve.StatsAccumulator.results

MiniObserve.Observation.@observe