StatsAccumulator.jl

MiniObserve.jl

Index

Observation.jl

StatsAccumulator.jl

StatsAccumulatorBase.jl

MiniObserve.StatsAccumulator.CountAcc — Type

Count number of true values in input. Results are returned as (;n).

MiniObserve.StatsAccumulator.HistAcc — Type

Track a histogram of input values. Results are returned as a (;bins::Vector{Int}).

MiniObserve.StatsAccumulator.HistAcc — Method

HistAcc()
HistAcc(min)
HistAcc(min, width)
HistAcc(min, width, max; count_below_min, count_above_max)

Construct a histogram accumulator.

Arguments:

min - minimum value [T(0)]

width - bin size [T(1)]

max - maximum value. Set lower than or equal to min to let the histogram adjust size automatically. [min]

count_below_min - whether values lower than min are ignored or counted in the first bin [false]

count_above_max - whether values larger than max are ignored or counted in the last bin [false]

MiniObserve.StatsAccumulator.MaxMinAcc — Type

Keep track of maximum and minimum of input. Results are returned as (;max, min).

MiniObserve.StatsAccumulator.MeanVarAcc — Type

Calculate mean and variance of input. Results are given as (;mean, var)

MiniObserve.StatsAccumulator.SumAcc — Type

Calculate sum of input. Results are returned as (;n).

Index

MiniObserve.Observation

MiniObserve.StatsAccumulatorBase

MiniObserve.StatsAccumulator.CountAcc

MiniObserve.StatsAccumulator.HistAcc

MiniObserve.StatsAccumulator.HistAcc

MiniObserve.StatsAccumulator.MaxMinAcc

MiniObserve.StatsAccumulator.MeanVarAcc

MiniObserve.StatsAccumulator.SumAcc

MiniObserve.Observation.joined_named_tuple_T

MiniObserve.Observation.log_results

MiniObserve.Observation.print_header

MiniObserve.Observation.stat_names

MiniObserve.Observation.to_named_tuple

MiniObserve.Observation.tuple_type

MiniObserve.StatsAccumulatorBase.result_type

MiniObserve.StatsAccumulatorBase.results

MiniObserve.Observation.@observe